Move json util and base64 library into a separate package
