+++ extensions/BUILD
@@ -38,8 +38,8 @@ wasm_cc_binary(
     ],
     copts = ["-UNULL_PLUGIN"],
     deps = [
-        "//extensions/common:json_util_wasm",
         "//extensions/common:node_info_fb_cc",
+        "//extensions/common/wasm:json_util",
         "//external:abseil_strings",
         "//external:abseil_time",
         "@proxy_wasm_cpp_sdk//:proxy_wasm_intrinsics",
@@ -56,14 +56,14 @@ wasm_cc_binary(
         "//extensions/common:proto_util.h",
         "//extensions/common:util.cc",
         "//extensions/common:util.h",
-        "//extensions/metadata_exchange:base64.h",
+        "//extensions/common/wasm:base64.h",
         "//extensions/metadata_exchange:plugin.cc",
         "//extensions/metadata_exchange:plugin.h",
     ],
     copts = ["-UNULL_PLUGIN"],
     deps = [
-        "//extensions/common:json_util_wasm",
         "//extensions/common:node_info_fb_cc",
+        "//extensions/common/wasm:json_util",
         "//extensions/metadata_exchange:declare_property_proto_cc",
         "//external:abseil_strings",
         "//external:abseil_time",
@@ -85,8 +85,8 @@ wasm_cc_binary(
     copts = ["-UNULL_PLUGIN"],
     deps = [
         "//extensions/attributegen:config_cc_proto",
-        "//extensions/common:json_util_wasm",
         "//extensions/common:node_info_fb_cc",
+        "//extensions/common/wasm:json_util",
         "//external:abseil_strings",
         "//external:abseil_time",
         "@proxy_wasm_cpp_sdk//:proxy_wasm_intrinsics_full",
@@ -103,12 +103,12 @@ wasm_cc_binary(
         "//extensions/common:context.h",
         "//extensions/common:util.cc",
         "//extensions/common:util.h",
-        "//extensions/metadata_exchange:base64.h",
+        "//extensions/common/wasm:base64.h",
     ],
     copts = ["-UNULL_PLUGIN"],
     deps = [
-        "//extensions/common:json_util_wasm",
         "//extensions/common:node_info_fb_cc",
+        "//extensions/common/wasm:json_util",
         "//external:abseil_strings",
         "@proxy_wasm_cpp_sdk//:proxy_wasm_intrinsics_full",
     ],
